+++ src/main/java/com/crpreparacoes/crpreparacoes/bodyrequestinput/bikePart/EditBikePartRequest.java
@@ -10,5 +10,5 @@ public class EditBikePartRequest {
     private String name;
     private Double value;
     private Integer stockQuantity;
-    private Integer bikeId;
+    private Long bikeId;
 }
